chore: Add new scenes for creating and editing bundles
This commit adds new scenes for creating and editing bundles in the Telegram bot. The scenes include "createBundle" and "editBundle" and are imported into the main "index" scene. This allows users to create new bundles with a name, price, description, and list of channels. It also enables editing existing bundles. The scenes are added to the "stage" object for scene management.